fix issue

+++ src/api/v1/modules/notifications/notification.service.ts
@@ -101,6 +101,8 @@ export default class NotificationService {
             comment: comment ? {id: comment.id} : null
         })
 
+        if(!notification) return
+
         await this.notificationRepository.remove(notification)
 
         return  notification
